Renamed es folder to src (more intuitive/expected)
