+++ src/dpt/message.js
